+++ lib/paper_trail/railtie.rb
@@ -1,0 +1,12 @@
+module PaperTrail
+  class Railtie < Rails::Railtie
+    config.paper_trail = ActiveSupport::OrderedOptions.new
+    initializer 'paper_trail.initialisation' do |app|
+      ActiveRecord::Base.send :include, PaperTrail::Model
+
+      unless app.config.paper_trail.fetch(:enabled, true)
+        PaperTrail.config.disable!
+      end
+    end
+  end
+end